Rename gui folder to assets.
